backend/requirements.txt
flask==3.0.3
PyJWT==2.8.0
bcrypt==4.1.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.32
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.1
statsmodels==0.14.2
python-dotenv==1.0.1
requests==2.32.3
python-dateutil==2.9.0
werkzeug==3.0.3
gunicorn==22.0.0
chardet==5.2.0
psycopg2-binary==2.9.9
reportlab==4.4.10
Flask-Limiter==3.5.0
